revert back index_t type in flash.h

+++ csrc/kernels/flash.h
@@ -16,7 +16,8 @@ constexpr int D_DIM = 2;
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 struct Qkv_params {
-    using index_t = int32_t;
+    using index_t = int64_t;
+
     // The QKV matrices.
     void *__restrict__ q_ptr;
     void *__restrict__ k_ptr;
